Refactored the GMDC theming system to allow for more extensibility
