+++ GroupMeClient.WpfUI/App.xaml
@@ -25,16 +25,8 @@
 
                 <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/Themes/Light.Blue.xaml" />
 
-                <ResourceDictionary Source="pack://application:,,,/Styles/GMDCGlobalResources.xaml" />
-                <ResourceDictionary Source="pack://application:,,,/Styles/GroupMeLight.xaml" />
-                
-                <ResourceDictionary Source="pack://application:,,,/Styles/GroupNameDropdown.xaml" />
-                <ResourceDictionary Source="pack://application:,,,/Styles/CopyableHyperlink.xaml" />
-                <ResourceDictionary Source="pack://application:,,,/Styles/MessageControlResources.xaml" />
-                <ResourceDictionary Source="pack://application:,,,/Styles/MarkdownStyles.xaml" />
-                <ResourceDictionary Source="pack://application:,,,/Styles/PopupDialog.xaml" />
-                <ResourceDictionary Source="pack://application:,,,/Styles/ScrollBar.xaml" />
-                <ResourceDictionary Source="pack://application:,,,/Styles/SendTextBox.xaml" />
+                <ResourceDictionary Source="pack://application:,,,/Resources/Resources.xaml" />
+                <ResourceDictionary Source="pack://application:,,,/Styles/Styles.xaml" />
                 
             </ResourceDictionary.MergedDictionaries>
