+++ GroupMeClient.WpfUI/Resources/GMDCGlobalResources.xaml
@@ -47,5 +47,6 @@
 
     <sys:Int32 x:Key="Zero">0</sys:Int32>
     <sys:Boolean x:Key="True">True</sys:Boolean>
+    <sys:Boolean x:Key="False">False</sys:Boolean>
 
 </ResourceDictionary>
