+++ GroupMeClient.WpfUI/Resources/Resources.xaml
@@ -1,0 +1,9 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="pack://application:,,,/Resources/GMDCGlobalResources.xaml" />
+        
+        <ResourceDictionary Source="pack://application:,,,/Resources/Themes/GMDC.Colors.Dark.xaml" />
+    </ResourceDictionary.MergedDictionaries>
+
+</ResourceDictionary>
