+++ GroupMeClient.WpfUI/Resources/Themes/GMDC.Colors.Dark.xaml
