+++ GroupMeClient.WpfUI/Resources/Themes/GMDC.Colors.Light.xaml
