+++ GroupMeClient.WpfUI/Styles/Styles.xaml
@@ -1,0 +1,12 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="pack://application:,,,/Styles/GroupNameDropdown.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/Styles/CopyableHyperlink.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/Styles/MessageControlResources.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/Styles/MarkdownStyles.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/Styles/PopupDialog.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/Styles/ScrollBar.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/Styles/SendTextBox.xaml" />
+    </ResourceDictionary.MergedDictionaries>
+</ResourceDictionary>
